o Implement /dev/graphics virtualizable access to 	  registers.  The actual context switch code is not 	  yet there.  But the rest of the magic (mapping/unmapping 	  the registers on demand is already in).
	o Interface for allowing binary-only console modules added.

	o My RRM bits, nothing really interesting now.

For now, I am assuming in the code that there will be a minor per
real graphics device, and that the X server will go and open
/dev/graphicsN instead of opening /dev/graphics and using the
gfx_attach_board->board variable.

The interface the X server uses is not clear.

I believe it will be pretty easy to provide the stripped down
shmiq interface.  Not only that, but we can even provide the
same ABI (yes, ABI) for SGI shmiq input modules (yep, looks
pretty easy once you have an strace that shows this information).

Now, the only thing missing is figuring what is wrong my current
conception of the use of shmiq.  I just can't get it.