Next iteration, that should be it [tm].
* offset.h now depends on include/linux/autoconf.h as suggested by Ulf.
That doesn't really make a difference, but it feels better :)
* If include/asm-$(ARCH)/offset.h doesn't exist during a "make dep",
create an empty one so that the dependecies are built correctly.
That seems to be the correct fix.
* Added a mrproper target to arch/$(ARCH)/tools/Makefile so that
"make dep clean vmlinux" works. include/asm-$(ARCH)/offset.h is now beeing
deleted with a "make mrproper".
* While beeing at it, I added a mrproper target in arch/$(ARCH)/boot/Makefile
as well.

I'd appreciate if someone with a working MIPS64 toolchain (hint, hint) could
check if this works as expected for MIPS64. Toggling CONFIG_SMP should rebuild
nearly the whole kernel without the need for a "make clean" or something like
that.