Redo R4600 workaround based on new information from PMC-Sierra.  This improves interrupt latency significantly and fixes another problem with the old workaround so should be considered a critical upgrade for R4600 systems.
Apply CPU workarounds only to those machines which were actually shipped
with the affected processors.  As for the R4600 the assumption is only
IP22 was shipped with V1.7 and V2.0 processors and SNI RM200C only with
V2.0 processors; for all other systems the workarounds are disabled.

Control workarounds based on value of the preprocessor symbol, not
based on it's definedness.

Move R5432 workaround to the (hopefully ...) right place and copy it
to those places where it was missing.