Consolidate all the MIPS PCI code in arch/mips/pci.
MIPS PCI code had started to share too many properties with a barbed wire
fence, so I decieded to move all the code into a single directory.  No
functional changes in the code so far.  Pretty much every piece of PCI
code we have was broken in interesting ways and code was duplicated
way too often, worst offender were systems based on the GT64120 and similar
system controllers.  Next step will be factoring out common code which
probably won't be too easy due to the amount of bastardization that has
happened to the code.  So policy is to cleanup, no prisoners taken.