IP22: Fix serial console detection
From: Kaj-Michael Lang <milang@tal.org>

In ip22-setup.c the checks for serial/graphics console logic does
not check if ARCS console=g but the machine is using serial console, as
it does if no keyboard is attached.

This patch adds a check if ConsoleOut is serial. There might also be
support for other graphics than Newport soon...

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>