MIPS: Fix oops in r4k_dma_cache_inv
The recent fix "MIPS: Avoid destructive invalidation on partial cachelines."
resulted potentially in an address error.

The reason for this is that there is no clearly documented alignment
requirement for the cache instruction on MIPS processors.  Some even
document that a cache instruction cannot result result in an address
error and some processors, among them the RM5200 and RM7000 QED
processors will throw a an address error for cache hit ops with
insufficient alignment - whatever the exact alignment requirement is.

Credits to Yoichi Yuasa <yoichi_yuasa@tripeaks.co.jp> for providing a
similar patch.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>