MIPS: Sibyte, SMTC: Use generic method for irq statistics
The structure 'struct kernel_stat' defines the 'irqs' array as its field
only when CONFIG_GENERIC_HARDIRQS is not set.  MIPS however sets
GENERIC_HARDIRQS for all platforms.  As the result, the following build
error is produced:

  CC      arch/mips/sibyte/bcm1480/smp.o
/home/ralf/src/linux/linux-queue/arch/mips/sibyte/bcm1480/smp.c: In function
‘bcm1480_mailbox_interrupt’:
/home/ralf/src/linux/linux-queue/arch/mips/sibyte/bcm1480/smp.c:183: error:
‘struct kernel_stat’ has no member named ‘irqs’

Similar for the other files modified by this patch.  This patch fixes the
build error by using the generic method for the irq statistics.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>