[loongson] change naming methods
to make source code of loongson sharable to the machines(such as gdium)
made by the other companies, we rename arch/mips/lemote to
arch/mips/loongson, asm/mach-lemote to asm/mach-loongson, and rename
lm2e to the name of the machine: fuloong-2e. accordingly, FULONG are
renamed to FULOONG2E to make it distinguishable to the future FULOONG2F.
and also, some other relative tuning is needed.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>