MIPS: AR7: Make board code register ar7_wdt as a platform device
This patch makes the board code register the ar7_wdt driver as a platform
device. We move the dynamic resource calculation here since the driver
should not be aware of the AR7 SoC version it is running on.

[This commit is based on commits d47fbb5996269f3081deb7a1b4ba717557256555
and 2ad3e4ae4aee996035f802c66065c93fec5607d2 from the master branch - Ralf]

Signed-off-by: Florian Fainelli <florian@openwrt.org>
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>