MIPS: add support for gzip/bzip2/lzma compressed kernel images
This patch will help to generate smaller kernel images for linux-MIPS,

Here is the effect when using lzma:

$ ls -sh vmlinux
7.1M vmlinux
$ ls -sh arch/mips/boot/compressed/vmlinuz
1.5M arch/mips/boot/compressed/vmlinuz

Have tested the 32bit kernel on Qemu/Malta and 64bit kernel on FuLoong
Mini PC. both of them work well.

and this revision incorporates the feedback from Alexander Clouter
<alex@digriz.org.uk>, he helped to test it on AR7[1] based Linksys
WAG54Gv2 and gave good suggestion on board-independence.

NOTE: this should work for the other MIPS-based machines, but I have
used the command bc in the Makefile to calculate the load address of the
compressed kernel. I'm not sure this is suitable.  perhaps I need to
rewrite this part in C program or somebody help to simplify the current
implementation.

TODO:
   1. add checksum support
   2. rewrite the code for calculating load address of compressed kernel

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>