Loongson: 2F: enable fixups of binutils
In function as-option, the parameter is passed to CC, without "-Wa,"
prefix. So the parameter should have "-Wa," prefix as shown in
script/Kbuild.include. Also because ,(comma) is as-option's parameter
delimeter, the comma in "-Wa," should be replaced by $(comma).

Documentation/kbuild/makefiles.txt:

as-option = $(call try-run,\
        $(CC) $(KBUILD_CFLAGS) $(1) -c -xassembler /dev/null -o "$$TMP",$(1),$(2))

and currently, there is no proof of need to use -mfix-ls2f-kernel for
non-loongson2f, so just enable it for loongson2f.

This option -mfix-ls2f-kernel is provided by the as patch from Lemote:
http://archlinux-for-loongson.googlegroups.com/web/binutils-2.19.1-loongson2f-3.patch

NOTE: The latest patch(for binutils 2.20) of this fixup is provided by
two new options: -mfix-loongson2f-nop and -mfix-loongson2f-jump, this
two will be sent to the mainline linux after the patch for binutils is
applied.

Signed-off-by: Zhang Le <r0bertz@gentoo.org>
Signed-off-by: Wu Zhangjin <wuzj@lemote.com>