[loongson] loongson2f: add basic pm support framework
This patch add basic pm(STR/Standby) support for loongson2f family
machines, loongson2f have a specific feature: when we set it's frequency
to ZERO, it will goto a wait mode, and then can be waked up by an
external interrupt. so, if we setup suitable interrupts before putting
it into wait mode, we will can wake it up whenever we want via sending
the relative interrupts to it.

These interrupts are board-specific, Yeeloong2F use the keyboard
interrupt and SCI interrupt, but LingLoong and Fuloong2F use the
interrupts connected to the processors directly. and BTW: some old
LingLoong and FuLoong2F have no such interrupts connected, so, there is
no way to wake them up from standby mode. and therefore, please do not
enable the kernel support for them.

The board-specific pm support will be added in the coming patches.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>