[loongson] mem.c: Register reserved memory pages
This patch will Fix Hiberation Error with FLATMEM on loongson family
machines.

when CONFIG_FLATMEM enabled, Hibernation will try to shrink the memory
pages not reserved and is valid checked by pfn_valid(). but we did not
register the reserved pages before. so, Hibernation tried to handle the
reserved pages which made itself fail.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>