MIPS: Add support of LZO-compressed kernels
This patch need the Compressed kernel support for MIPS and the support
of LZO-compressed kernel, both of them are not available in the official
kernel yet, but you can get them in the following links respectively:

[PATCH -v2] MIPS: add support for gzip/bzip2/lzma compressed kernel image

http://www.linux-mips.org/archives/linux-mips/2009-10/msg00110.html

[PATCH v2.1 2/5] Add support for LZO-compressed kernels
http://marc.info/?l=linux-kernel&m=125525500919174&w=2

The advantage of this algorithm:

Uncompressed size: 3.24Mo
gzip  1.61Mo 0.72s
lzo   1.75Mo 0.48s

and what about the compression ratio?

$ ls -sh vmlinux  (original)
6.0M vmlinux
$ ls -sh vmlinuz (lzma)
1.2M vmlinuz
$ ls -sh vmlinuz (bzip2)
1.5M vmlinuz
$ ls -sh vmlinuz (gzip)
1.6M vmlinuz
$ ls -sh vmlinuz (lzop)
1.7M vmlinuz

the time for decompressing kernels compressed by gzip,bzip2,lzma:

	SIZE	AVG. TIME(in 4 times)
gzip   2011314       2.64
bzip2  1831170       4.0625
lzma   1470938       3.2075

so, if you want to get the best compression ratio, you are recommended to use
LZMA, and if you need the fastest one, LZO is recommanded.

and BTW: if you need to use the gzip,bzip2,lzma,lzo support, the
corresponding tools are need to install in your machine to compress the
original kernel, they are gzip,bzip2,lzma and lzop.

(Thanks very much to Alexander Clouter <alex@digriz.org.uk> for giving the
information about LZO to me)

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>