[loongson] yeeloong2f, rtl8187b: a few fixups
This patch simplify the controlling logic of rtl8187b and speedup the
action of led light with Fn+F5.

If we power on the wifi, power on it whenever the old status it is,
which will avoid the possible un-sync problems, and also, we just turn
on the led whenever the rtl8187b is up or not. we just want to tell the
user, we have powered it on, if it not work, the NetworkManager will not
work and then you get the real status of the wifi.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>