[loongson] yeeloong2f, backlight: fixes the problem of Fn+F2
If with the latest EC(PQ1D26 >= version), after pressing Fn + F2, EC
will does corresponding action for us(turn off/on the screen). and if we
do the actions in kernel again, the brightness will can not resume to
the original level, but become to the lowest level.

So, Here we need to check the version of EC, if it is older than PQ1D26,
we do the actions in kernel, otherwise, ignore it.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>