Revert "[loongson] yeeloong2f,rtl8187b: cleanup the rfkill support"
Seems we really need the Fn + F5 as a switch, for we can not ensure the
status in the REG_WLAN is the real current status of the wifi.

and also, 'Cause if enable the wifi by default, it will bring us with
significant performance influence and some places need the wifi disabled
by default, such as in the airplane.

So, we keep back what we have designed before.

And of course, you can enable it when booting via add this line into
/etc/rc.local:

echo 1 > /sys/class/rfkill/rfkill0/state

And hope the kpowersave or gnome-power-manager of nm-applet provide such
an interface for helping setting the default booting status of the wifi.
and also give a choice for recording the status of the last Fn + F5 or
not.

This reverts commit efe984d8afacbedbe127080a6934b2e934e2be63.