[loongson] yeeloong2f: add LID open event as the wakeup event
Yeeloong2F netbook has an embedded controller(kb3310b) to manage the LID
action. when the LID is closed/opened, a SCI interrupt is sent out and
the corresponding event is saved to an EC register for later query.

This patch allow the LID open action to wake up the processor from wait
mode if it is in the suspend mode.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>