[loongson] MTD: add ROM of PMON flushing/dumping support
This is a driver for flushing/dumping the bootloader(pmon) of
fuloong2e,fuloong2f,yeeloong2f in Linux. which is provided to the
developers, but not for the generic users.

NOTE: It's very _dangerous_ to do the flush the PMON, if unsure, please
NOT do it, otherwise, you will get a break your machine.

To enable this option, two kernel options are needed:

MTD
MTD_PARTITIONS

and also, you need to compile this driver yourselves, it provides no way
to compile it.

Here is a basic usage:

1. load necessary modules

$ modprobe mtd
$ modprobe mtdblock
$ modprobe yeeloong_pmon

2. read the flash

$ dd if=/dev/mtdblock0 of=pmon.bin

3. write the falsh

$ dd if=pmon.bin of=/dev/mtdblock0

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>