MIPS: EARLY_PRINTK: Fixup of dependency
In the old version, if the platform select SYS_HAS_EARLY_PRINTK and
users not select DEBUG_KERNEL, there is no interface for user to enable
or disable the EARLY_PRINTK option, it will be enabled all the time.
this will waste a little bit of memory and slow down the booting of
kernel but the users never know it.

This patch will only enable that option when the DEBUG_KERNEL is
enabled.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>