Loongson: YeeLoong2F: add a new rtl8187b wifi driver
Changes:

	o 2009/12/18: select WIRELESS_EXT and WEXT_PRIV

There is an existing rtl8187 wifi driver in linux, but the performance
is not good on yeeloong2f. so, we add a new one from realtek, it is
rtl8187B_linux_26.1051.0116.2009_release. before, it stay out of the
kernel, which is not convenient to the users and linux distribution
developers, so I merge it in to the linux kernel.

To make it work on linux-2.6.30 and later versions, there is a need to
add the new net_device_ops APIs, and remove one of the duplicated
function.

and also, to ensure it works well with the standard NetworkManager,
rfkill support is added.

By default, the wifi is powered off, if you want to enable it when the
system is booting, add this line into your /etc/rc.local:

echo 1 > /sys/class/rfkill/rfkill0/state

NOTE! this driver is not suitable to upstream, should only keep in this
git repository currently.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>