Loongson: YeeLoong: Rtl8187b: Fixup of rfkill status report
Sam Geeraerts <samgee@elmundolibre.be> have reported:

"My wifi doesn't come back when coming out of suspend. NetworkManager
says that network is disabled and doesn't list any interfaces. rtl8187b
still shows up in lsmod.My wifi doesn't come back when coming out of
suspend. NetworkManager says that network is disabled and doesn't list
any interfaces. rtl8187b still shows up in lsmod."

This is a "weird" action for we have reported the blocked status into
user-space when suspending but this time, the user-space will not report
it for the applications have been freezed when suspending, the report
will be delayed to resuming, this patch will fix it!

Reported-by: Sam Geeraerts <samgee@elmundolibre.be>
Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>