MIPS: MIPS: fix vmlinuz build when only 32bit math shell is available (cont.)
The old revision not works with dash, this patch fixes it with the
support of expr command.

This basic method is the same, but to ensure it work with dash, we can
not use the bash specific syntax to calculate, to make it portable, we
use the expr command instead.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>