Loongson: Cleanup the halt and poweroff action
In the old source code, I have let halt and poweroff do the same action,
   but in reality, they have different meanings.

As the manpage of shutdown shows:

-r     Reboot after shutdown.
-H     Halt action is to halt or drop into boot monitor on systems that support it.
-P     Halt action is to turn off the power.

and in the real world, some machines did not provide a power button, and
the shutdown will work as reset, so, we need to provide a mechanism to
let the users turn off the power safely without breaking the system,
such a mechanism is "halt".

"halt" only put the system into a dead loop or a power-save mode, and
print some information to the screen to tell the users to turn off the
power safely.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>