Loongson: Oprofile: Enable it when CONFIG_OPROFILE=m
When the oprofile is compiled as a module, the do_IRQ() is not called in
arch/mips/loongson/lemote-2f/irq.c for the wrong #ifdef there.

This patch fixes it via calling do_IRQ() whenever oprofile is compiled
as a module(CONFIG_OPROFILE_MODULE is defined) or compiled into the
kernel(CONFIG_OPROFILE is defined).

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>