Loongson: Disable early_printk by default in defconfig
For the serial port address in some machines are different from the
config in arch/mips/loongson/common/uart_base.c and will make the kernel
fail on booting when trying to access the wrong serial port.

This patch just deselect the EARLY_PRINTK to not enable early printk.
and in reality, early printk is only needed when debugging, disabling it
make also speedup the booting of kernel.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>