MIPS: Loongson: Fix the hang introduced by loongson2_wait()
loongson2_wait() put loongson into standby mode and must be waked up by
the external interrupts.

r4k timer works in nohz/oneshot mode, no interrupt will be sent to wake
the loongson from the standby mode if loongson is in standby mode and
the system will hang. but the external mfgpt timer, which has lower
resolution, can not work in the real oneshot mode and will send the
interrupt periodically to wake up the loongson from standby mode, no
hange take place.

BTW, to make the cpufreq of loongson works normally, we'd better use a
high resloution timer, otherwise, it will not function normally then the
power consumption will not saved. currently, there is only one high
resolution timer available, that is r4k timer, but unfortunately, for
the freq of r4k timer is related to the cpu freq, if cpufreq is enabled,
the timer system will be broken. To solve this problem, we may need to
find a solution to 'recover' the timer system even if cpufreq is
enabled, or the Loongson designer provide another freq-fixed timer.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>