MIPS: Loongson: Fixup of the slowdown problem
The commit "MIPS: Loongson: Clean up the PM support" have tried to
improve the PM support for Loongson but slowed down the machine.

The problem is it added a default loongson2_wait() with putting the
processor to the wait mode which can only be waked up by the external
interrupts and at the same time, it changed the HZ to a smaller value to
reduce the timer interrupts, this two actions means the processors will
be always in idle mode only if the timer interrupts wake it up. For the
HZ is too smaller, the whole machine is slowed down.

This patch fix it via only setting the cpu into the 2nd lowest level in
loongson2_wait(), then, even there is no interrupt, the processor runs
and can then be changed to a higher level if there are lots of works to
do. and to ensure the tasks can be scheduled timely, we change the HZ
back to a higher value: 250HZ.

Reported-by: Liu Shiwei <liushiwei@gmail.com>
Reported-by: Alexandre Oliva <lxoliva@fsfla.org>
Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>