MIPS: Loongson: Add MFGPT based hrtimer implementation
This patch is an effort to implement a high-res timer based on cs5536-mfgpt
hardware. Many thanks to Zhangjin for his help, and his original idea.

The patch is against the 2.6.35+ tree. Given the patch, the kernel can switch
to the hrtimer mode, hence run under the tickless mode. The cpufreq module now
works: CPU frequency will be throttled down at idle as expected. The CPU sleeps
as well, but we still need to check the how well the sleep goes.

Given all these nice things, there are still some issues remaining.

First and foremost, the system is not stable enough: I do experience unexpected
hangs sometimes. Maybe some interactions between hrtimer and the system, or
just the timer's fault. More testing needs to be done. My experience is that
the higher frequency we choose for clock_event_device (mfgpt0), the stabler of
the system. On the other hand, the less benefit we can get from a tickless
system: for example, if we use 14M for mfgpt0, the maximum duration the timer
can provide is just slight more than 1 tick. Because the timer has to fire to
maintain the system beat, there is no point to switch to tickless mode: we end
up with about 250 interrupts anyway.

Second, the clocksource is implemented on based on mfgpt1. Its frequency is set
as 16K. Due to the frequency disparity between mfgpt0 (high freq) and mfgpt1
(low freq), we see many interrupt retries.

The current setting chosen in the patch is a trade-off given all these factors.

Third, the mfgpt_base is variable so we have to poll its value inside the
interrupt handler. I found that its value only changes once at boot time when
the cs5536 pci is set up. Somehow I believe this is because PMON and the
current kernel source use different settings. Not sure whether it is easy to
align them. Dealing with it at the timer module is quite painful.

Signed-off-by: Gang Liang <randomizedthinking@gmail.com>
[ Fixes the errors and warnings reported by scripts/checkpatch.pl ]
Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>