MIPS: YeeLoong: Tune the relative capacity of the battery driver
The old driver has reported a wrong value for the relative capacity and
made the gnome power manager can not do related actions timely before
the hardware is turned off forcely.

When the capacity is lower than about 2%, the the hardware will be
turned off forcely without any protection. To avoid such a situation, we
must tell a smaller capacity(base on the value read from the register,
here we choose 3% to ensure the software have enough time to do actions)
to gnome power manager, then, it can do related actions earlier before
the laptop run out of the power.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>