MIPS: Loongson: Add a small defconfig for Lemote-2f machines
This patch adds a new small defconfig for a more functional support to
the Lemote-2F machines.

It is based on the minimal defconfig, but enables more functions:

  o sound
  o wireless
  o power management
  o platform support

All of them has been enabled as modules, so, even if you forget to copy
the modules, the kernel can boot normally with basic functions.

Now, we have 3 different defconfig for Lemote-2F machines.

  o minimal_defconfig

  Suitable for a server and resuce/recover system and also for (as a
  base) speeding up the develpment of a new kernel feature

  o small_defconfig

  Suitable for a laptop and for kernel maintaining/updating

  o defconfig

  Suitable for a full linux distruction

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>